GRIDEDIT-265: change to size_t  m_boundaryEdgesCache, m_nodeTopologyMapping, m_numTopologyNodes and m_numTopologyFaces in Smoother

+++ include/MeshKernel/FlipEdges.hpp
@@ -72,7 +72,7 @@ namespace meshkernel
         /// @brief Compute the difference with the optimal number of edges by counting the numbers of edges that
         ///        connect nodes firstNode and secondNode, and are on the land boundary path (comp_nnow)
         /// @returns Difference form optimum
-        [[nodiscard]] size_t DifferenceFromOptimum(size_t nodeIndex, size_t firstNode, size_t secondNode) const;
+        [[nodiscard]] int DifferenceFromOptimum(size_t nodeIndex, size_t firstNode, size_t secondNode) const;
 
         /// @brief Remove a connected edge from a node
         /// @param[in] edgeIndex The index of the edge to remove
